Separate HeaderParser module
