+++ lib/serum/header_parser.ex
@@ -20,12 +20,11 @@ defmodule Serum.HeaderParser do
   actual value of a metadata.
   """
 
-  @date_format1 "{YYYY}-{0M}-{0D} {h24}:{m}:{s}"
-  @date_format2 "{YYYY}-{0M}-{0D}"
+  alias Serum.HeaderParser.ValueTransformer
 
   @type options :: [{atom, value_type}]
   @type value_type :: :string | :integer | :datetime | {:list, value_type}
-  @type value :: binary | integer | [binary] | [integer]
+  @type value :: binary | integer | DateTime.t() | [binary] | [integer] | [DateTime.t()]
   @type parse_result :: {:ok, {map(), binary()}} | {:invalid, binary()}
 
   @typep extract_ok :: {:ok, [binary], binary}
@@ -168,65 +167,9 @@ defmodule Serum.HeaderParser do
   defp transform_values([{k, v} | rest], options, acc) do
     atom_k = String.to_existing_atom(k)
 
-    case transform_value(k, String.trim(v), options[atom_k]) do
+    case ValueTransformer.transform_value(k, String.trim(v), options[atom_k]) do
       {:error, _} = error -> error
       value -> transform_values(rest, options, [{atom_k, value} | acc])
     end
   end
-
-  @spec transform_value(binary, binary, value_type) :: value | {:error, binary}
-
-  defp transform_value(_key, valstr, :string) do
-    valstr
-  end
-
-  defp transform_value(key, valstr, :integer) do
-    case Integer.parse(valstr) do
-      {value, ""} -> value
-      _ -> {:error, "`#{key}`: invalid integer"}
-    end
-  end
-
-  defp transform_value(key, valstr, :datetime) do
-    case Timex.parse(valstr, @date_format1) do
-      {:ok, dt} ->
-        dt |> Timex.to_erl() |> Timex.to_datetime(:local)
-
-      {:error, _msg} ->
-        case Timex.parse(valstr, @date_format2) do
-          {:ok, dt} ->
-            dt |> Timex.to_erl() |> Timex.to_datetime(:local)
-
-          {:error, msg} ->
-            {:error, "`#{key}`: " <> msg}
-        end
-    end
-  end
-
-  defp transform_value(key, _valstr, {:list, {:list, _type}}) do
-    {:error, "`#{key}`: \"list of lists\" type is not supported"}
-  end
-
-  defp transform_value(key, valstr, {:list, type}) when is_atom(type) do
-    list =
-      valstr
-      |> String.split(",")
-      |> Stream.map(&String.trim/1)
-      |> Stream.reject(&(&1 == ""))
-      |> Stream.map(&transform_value(key, &1, type))
-
-    case Enum.filter(list, &error?/1) do
-      [] -> Enum.to_list(list)
-      [{:error, _} = error | _] -> error
-    end
-  end
-
-  defp transform_value(key, _valstr, _type) do
-    {:error, "`#{key}`: invalid value type"}
-  end
-
-  @spec error?(term) :: boolean
-
-  defp error?({:error, _}), do: true
-  defp error?(_), do: false
 end
